requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langgraph==0.2.28
pypdf==4.3.1
pdfplumber==0.11.4
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
google-genai>=0.8.0
